+++ src/collective/volto/formsupport/captcha/norobots.py
@@ -1,5 +1,5 @@
-import json
-
+from . import CaptchaSupport
+from collective.volto.formsupport import _
 from collective.z3cform.norobots.browser.interfaces import INorobotsWidgetSettings
 from plone import api
 from plone.registry.interfaces import IRegistry
@@ -7,9 +7,7 @@ from zExceptions import BadRequest
 from zope.component import queryUtility
 from zope.i18n import translate
 
-from collective.volto.formsupport import _
-
-from . import CaptchaSupport
+import json
 
 
 class NoRobotsSupport(CaptchaSupport):