+++ src/collective/volto/formsupport/restapi/services/submit_form/post.py
@@ -9,7 +9,9 @@ from copy import deepcopy
 from datetime import datetime
 from email import policy
 from email.message import EmailMessage
+from io import BytesIO
 from plone import api
+from plone.base.interfaces.controlpanel import IMailSchema
 from plone.protect.interfaces import IDisableCSRFProtection
 from plone.registry.interfaces import IRegistry
 from plone.restapi.deserializer import json_body
@@ -40,7 +42,7 @@ CTE = os.environ.get("MAIL_CONTENT_TRANSFER_ENCODING", None)
 
 
 @implementer(IPostEvent)
-class PostEventService(object):
+class PostEventService:
     def __init__(self, context, data):
         self.context = context
         self.data = data
@@ -48,7 +50,7 @@ class PostEventService(object):
 
 class SubmitPost(Service):
     def __init__(self, context, request):
-        super(SubmitPost, self).__init__(context, request)
+        super().__init__(context, request)
 
         self.block = {}
         self.form_data = self.cleanup_data()
@@ -77,7 +79,7 @@ class SubmitPost(Service):
                 message = translate(
                     _(
                         "mail_send_exception",
-                        default="Unable to send confirm email. Please retry later or contact site administator.",
+                        default="Unable to send confirm email. Please retry later or contact site administrator.",
                     ),
                     context=self.request,
                 )
@@ -218,7 +220,7 @@ class SubmitPost(Service):
             i = int(math.floor(math.log(attachments_len, 1024)))
             p = math.pow(1024, i)
             s = round(attachments_len / p, 2)
-            uploaded_str = "{} {}".format(s, size_name[i])
+            uploaded_str = f"{s} {size_name[i]}"
             raise BadRequest(
                 translate(
                     _(
@@ -468,11 +470,11 @@ class SubmitPost(Service):
                     continue
                 content_type = value.get("content-type", content_type)
                 filename = value.get("filename", filename)
-                if isinstance(file_data, six.text_type):
+                if isinstance(file_data, str):
                     file_data = file_data.encode("utf-8")
                 if "encoding" in value:
                     file_data = codecs.decode(file_data, value["encoding"])
-                if isinstance(file_data, six.text_type):
+                if isinstance(file_data, str):
                     file_data = file_data.encode("utf-8")
             else:
                 file_data = value
@@ -492,7 +494,7 @@ class SubmitPost(Service):
             .replace(":", "")
         )
         filename = f"formdata_{now}.xml"
-        output = six.BytesIO()
+        output = BytesIO()
         xmlRoot = Element("form")
 
         for field in self.filter_parameters():