+++ src/collective/volto/formsupport/scripts/cleansing.py
@@ -1,14 +1,12 @@
-# -*- coding: utf-8 -*-
-import sys
-
-import click
-import transaction
+from collective.volto.formsupport.interfaces import IFormDataStore
+from collective.volto.formsupport.restapi.services.form_data.form_data import FormData
 from plone import api
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
 
-from collective.volto.formsupport.interfaces import IFormDataStore
-from collective.volto.formsupport.restapi.services.form_data.form_data import FormData
+import click
+import sys
+import transaction
 
 
 @click.command(